2. Multi-Language Invoice Project/requirements.txt
streamlit
google-generativeai
langchain
PyPDF
chromadb
python-dotenv